libopalevents
This library contains code to parse and print OPAL events. Several projects
have a version or other of this code and the level of duplication has become
unmaintainable.

This library plans to reduce the workload and keep all the problems in one
place. Projects parsing OPAL events should consider using this library.